requirements.txt
absl-py
astor
astunparse
attrs
backcall
bleach
cachetools
certifi
chardet
colorama
cycler
decorator
defusedxml
entrypoints
gast
google-auth
google-auth-oauthlib
google-pasta
graphviz
grpcio
h5py
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
jedi
Jinja2
json5
jsonschema
jupyter-client
jupyter-core
jupyterlab
jupyterlab-server
Keras-Applications
Keras-Preprocessing
kiwisolver
Markdown
MarkupSafe
matplotlib
mistune
nbconvert
nbformat
notebook
numpy
oauthlib
opencv-python
opt-einsum
packaging
pandocfilters
parso
pickleshare
Pillow
prometheus-client
prompt-toolkit
protobuf
pyasn1
pyasn1-modules
pydot
Pygments
pyparsing
pyrsistent
python-dateutil
pyzmq
requests
requests-oauthlib
rsa
scipy
Send2Trash
six
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
terminado
testpath
tornado
traitlets
urllib3
wcwidth
webencodings
Werkzeug
wrapt
zipp
